Refactor certifier tests
